% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_data.R
\name{compare_mixing_proportions}
\alias{compare_mixing_proportions}
\title{Compare Mixing Proportions Between Data Streams}
\usage{
compare_mixing_proportions(synth_list_d, synth_list_c, mu_tab)
}
\arguments{
\item{synth_list_d}{A list representing the synthetic dataset with divergent mixing proportions.}

\item{synth_list_c}{A list representing the synthetic dataset with convergent mixing proportions.}

\item{mu_tab}{A data frame containing the mean tracer
signatures for each source. The first column should be named \code{source}, and
the remaining columns should contain numeric tracer values.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object visualizing the comparison of mixing proportions between the two data streams.
}
\description{
This function generates a plot to compare the mixing proportions between two data streams
(e.g., chemical tracers and community composition data) for synthetic datasets with agreement and disagreement.
}
\details{
The function compares the mixing proportions from two data streams (\code{stream_1_props} and \code{df_stream_2})
for both divergent and convergent synthetic datasets. It reshapes and aligns the data based on the
reference order provided in \code{mu_tab}, and then creates a scatter plot with points representing
the proportions from each data stream. The plot includes facets for each source and highlights
agreement or disagreement between the datasets.
}
\examples{
library(MixMustR)
compare_mixing_proportions(
  synth_list_d = synthetic_df_divergent,
  synth_list_c = synthetic_df_convergent,
  mu_tab = tracer_parameters$mus
)

}
\seealso{
\code{\link{synthetic_df_divergent}}, \code{\link{synthetic_df_convergent}}
}
