% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixMustR-package.R
\docType{package}
\name{MixMustR-package}
\alias{MixMustR}
\alias{MixMustR-package}
\alias{mixmustr}
\title{The MixMustR package.}
\description{
This is a flexible Bayesian mixture model package written in
the probabilistic programming language Stan (\url{https://mc-stan.org/}) for R
(\url{https://www.r-project.org/}). It estimates source mixing proportions by
incorporating simultaneous likelihood evaluation from two independent data
streams collected from the mixture of interest: one obtained from chemical
tracers/biomarkers (i.e., a single tracer measurement per observation, e.g.,
from stable isotopes and fatty acids), and another yielding source
composition (e.g., based on eDNA metabarcoding, particle-tracking
simulations, or other community-composition data). \code{MixMustR} also allows
for the estimation of an additional, unsampled source component to partially
relax the assumption that the mixing proportions from all samples sources
should sum up to 1. \code{MixMustR} should have wide applicability in ecological
studies, particularly given the growing usage and availability of multiple
tracers spanning traditional stable isotopes and eDNA to understand carbon
source-sink dynamics, and a mixture of stables isotope, fatty acids and eDNA
to unravel trophic interactions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://open-AIMS.github.io/MixMustR/}
  \item Report bugs at \url{https://github.com/open-AIMS/MixMustR/issues}
}

}
\author{
\strong{Maintainer}: Diego R. Barneche \email{d.barneche@aims.gov.au} (\href{https://orcid.org/0000-0002-4568-2362}{ORCID})

Authors:
\itemize{
  \item Julie Vercelloni \email{j.vercelloni@aims.gov.au} (\href{https://orcid.org/0000-0001-5227-014X}{ORCID})
  \item Edgar Santos-Fernández \email{edgar.santosfernandez@qut.edu.au} (\href{https://orcid.org/0000-0001-5962-5417}{ORCID})
  \item Kay L. Davis \email{ka.davis@aims.gov.au} (\href{https://orcid.org/0000-0002-7223-4303}{ORCID})
  \item Shaun K. Wilson \email{sk.wilson@aims.gov.au} (\href{https://orcid.org/0000-0002-4590-0948}{ORCID})
  \item Luke Thomas \email{l.thomas@aims.gov.au}
  \item James M. McGree \email{james.mcgree@qut.edu.au} (\href{https://orcid.org/0000-0003-2997-8929}{ORCID})
  \item Kerrie Mengersen \email{k.mengersen@qut.edu.au} (\href{https://orcid.org/0000-0001-8625-9168}{ORCID})
  \item Christopher J. Fulton \email{c.fulton@aims.gov.au} (\href{https://orcid.org/0000-0002-1140-1999}{ORCID})
}

}
