% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmustr_models.R
\docType{data}
\name{mixmustr_models}
\alias{mixmustr_models}
\title{MixMustR Models Configuration}
\format{
A data frame with 8 rows and 4 variables:
\describe{
\item{sample_tracer}{A logical vector, defaults to FALSE.
Should the model estimate uncertainty around sampled sources signatures?}
\item{fix_unsampled}{A logical vector, defaults to FALSE.
Should the model estimate uncertainty around the unsampled source
signatures or should it be fixed to mean across all sampled sources?}
\item{hierarchical}{A logical vector, defaults to FALSE.
Should all observations be treated as independent or should the model include
a hierarchical grouping structure?}
\item{code_path}{A character vector indicating the file path to the
corresponding Stan model file, based on the parameter combination.}
}
}
\source{
The data frame is generated in the \code{data-raw/mixmustr_models.R} script.
}
\usage{
mixmustr_models
}
\description{
This data frame defines the configuration of models used in the \code{MixMustR}
package. It specifies combinations of key parameters for generating and
running Stan models, which are fundamental to the functions
\code{mixmustr_wrangle_input} and \code{run_mixmustr_models}.
}
\details{
The \code{mixmustr_models} data frame is programmatically generated using all
possible combinations of the \code{sample_tracer}, \code{fix_unsampled}, and
\code{hierarchical} parameters. Each combination corresponds to a specific Stan
model file, whose path is stored in the \code{code_path} column.

This data frame is used internally by the \code{MixMustR} package to determine
which Stan model to use for a given analysis, based on user-specified
options.
}
\examples{
library(MixMustR)
data(mixmustr_models)
head(mixmustr_models)

# Access the Stan model path for a specific configuration
mixmustr_models$code_path[mixmustr_models$sample_tracer & mixmustr_models$hierarchical]
}
\keyword{datasets}
