% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{run_mixmod}
\alias{run_mixmod}
\title{Wrapper to run mixture model in Stan}
\usage{
run_mixmod(model_path, mod_name_suffix, ...)
}
\arguments{
\item{model_path}{Character scalar. Path to the \code{.stan} file to compile and
fit, typically generated by \code{\link{build_stancode}}.}

\item{mod_name_suffix}{A character string used to create a unique model name.}

\item{...}{Additional arguments passed to \code{\link[rstan]{stan}}, such as
\code{data}, \code{iter}, \code{chains}, or \code{control}.}
}
\value{
A fitted \code{stanfit} object, as returned by \code{\link[rstan]{stan}}.
}
\description{
Wrapper to run mixture model in Stan
}
\seealso{
\code{\link{run_mixmustr_models}}
}
