% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmustr-package.R
\docType{package}
\name{mixmustr-package}
\alias{mixmustr-package}
\alias{mixmustr}
\title{The 'mixmustr' package.}
\description{
This is a flexible Bayesian mixture model package written in
the probabilistic programming language Stan (\url{https://mc-stan.org/}) for R
(\url{https://www.r-project.org/}). It estimates source mixing proportions by
incorporating simultaneous likelihood evaluation from two independent data
streams collected from the mixture of interest: one obtained from chemical
tracers/biomarkers (i.e., a single tracer measurement per observation, e.g.,
from stable isotopes and fatty acids), and another yielding source
composition (e.g., based on eDNA). \code{mixmustr} also allows for the estimation
of an additional, unsampled source component to partially relax the
assumption that the mixing proportions from all samples sources should sum
up to 1. \code{mixmustr} should have wide applicability in ecological studies,
particularly given the growing usage and availability of multiple tracers
spanning traditional stable isotopes and eDNA to understand carbon
source-sink dynamics, and a mixture of stables isotope, fatty acids and eDNA
to unravel trophic interactions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://open-aims.github.io/mixmustr/}
  \item Report bugs at \url{https://github.com/open-aims/mixmustr/issues}
}

}
\author{
\strong{Maintainer}: Diego R. Barneche \email{d.barneche@aims.gov.au} (\href{https://orcid.org/0000-0002-4568-2362}{ORCID})

Authors:
\itemize{
  \item Chritopher J. Fulton \email{c.fulton@aims.gov.au} (\href{https://orcid.org/0000-0002-1140-1999}{ORCID})
}

}
