% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{mixmustr_bayes_R2}
\alias{mixmustr_bayes_R2}
\title{Compute a Bayesian version of R-squared for mixture models}
\usage{
mixmustr_bayes_R2(modfit, summary = TRUE, ...)
}
\arguments{
\item{modfit}{A fitted Stan model object containing posterior draws.
Typically produced by function \code{\link{run_mixmustr_models}}.}

\item{summary}{Should summary statistics be returned
instead of the raw values? Default is \code{TRUE}.}

\item{...}{Additional unused arguments.}
}
\value{
If \code{summary = TRUE}, a (J + 1) x 4 matrix is returned
(J + 1 = number of sampled + 1 unsampled source). Columns contain source
names and summary statistics (mean +/- 95\% HDI) of the Bayesian R-squared
values. If \code{summary = FALSE}, the posterior draws of the Bayesian
R-squared values are returned in an S x (J + 1) matrix (S is the number of
draws).
}
\description{
Compute a Bayesian version of R-squared for mixture models
}
\details{
For an introduction to the approach, see Gelman et al. (2019)
and \url{https://github.com/jgabry/bayes_R2/}.
}
\examples{
\dontrun{
library(mixmustr)
# mixmustr_models[6, ] runs quickest
model_fits <- run_mixmustr_models(
  mixmustr_models[6, ], synthetic_df_convergent, tracer_parameters,
  sigma_ln_rho = 0.1, iter = 1e4, warmup = 5e3, chains = 4, cores = 4
)
# Now compute R2, using data stream 2 as point of comparison
mixmustr_bayes_R2(
  model_fits[[1]]$model, data_streams_list = synthetic_df_convergent,
  target = "df_stream_2", order_ref = tracer_parameters$mus$source
)
}

}
\references{
Andrew Gelman, Ben Goodrich, Jonah Gabry & Aki Vehtari. (2019).
R-squared for Bayesian regression models, \emph{The American Statistician},
73(3):307-309. \code{10.1080/00031305.2018.1549100}
}
