% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{mixmustr_wrangle_input}
\alias{mixmustr_wrangle_input}
\title{mixmustr_wrangle_input}
\usage{
mixmustr_wrangle_input(
  data_streams_list,
  mu_tab,
  sig_tab = NULL,
  model_path,
  sigma_ln_rho,
  m,
  sample_tracer,
  fix_unsampled,
  hierarchical,
  ...
)
}
\arguments{
\item{data_streams_list}{A list containing the input
data streams for the model. It should include two
data frames named \code{df_stream_1} and\code{df_stream_2}. See
details for the expected structure of these.}

\item{mu_tab}{A data frame containing the mean tracer
signatures for each source. The first column should be named \code{source}, and
the remaining columns should contain numeric tracer values.}

\item{sig_tab}{(Optional) A data frame containing the
standard deviations of tracer signatures for each source. The structure must
match \code{mu_tab}. Required if \code{sample_tracer} is \code{TRUE} or if \code{fix_unsampled}
is \code{FALSE} and \code{sample_tracer} is \code{FALSE}.}

\item{model_path}{A character string specifying the path to the Stan model
file.}

\item{sigma_ln_rho}{A numeric value or matrix specifying the confidence
around the log mixing proportions. If a matrix, it must be of dimensions
N X J, with N being the number of observations and J being the number of
sources.}

\item{m}{A numeric value specifying the sample size for each source. Used
when \code{sample_tracer} is \code{TRUE}.}

\item{sample_tracer}{A logical vector, defaults to FALSE.}

\item{fix_unsampled}{A logical vector, defaults to FALSE.}

\item{hierarchical}{A logical vector, defaults to FALSE.
Should all observations be treated as independent or should the model include
a hierarchical grouping structure?}

\item{...}{Additional unused arguments.}
}
\value{
A list containing all the data elements necessary to run the Stan
code contained within \code{model_path}.
}
\description{
Creates the input list for the Stan model.
}
\details{
The \code{mixmustr_wrangle_input} function prepares the input data for the Stan
model:
\itemize{
\item \code{data_streams_list}: A list containing two data frames:
\itemize{
\item \code{df_stream_1}: Contains the observed data. If this is a hierarchical
dataset, the first column should be named \code{group}, representing the group
labels, and the remaining columns should contain numeric tracer values.
\item \code{df_stream_2}: Contains the log mixing proportions. It must have the same
number of rows (\code{N}) as \code{df_stream_1}. If the design is hierarchical, the
first column should be named \code{group}, matching the \code{group} column in
\code{df_stream_1}. The remaining columns (\code{source1}, \code{source2}, etc.) must
contain proportions (i.e., row sums up to 1, and no negative values).
}
\item \code{mu_tab}: A data frame containing the mean tracer signatures for each
source. The first column should be named \code{source}, and the remaining columns
should contain numeric tracer values, with names matching those in
data_streams_list$df_stream_1.
\item \code{sig_tab}: A data frame containing the standard deviations of tracer
signatures for each source. The structure must match that of \code{mu_tab}. It
cannot contain NAs.
\item \code{model_path}: A character string specifying the path to the Stan model
file.
\item \code{sigma_ln_rho}: A numeric value or matrix specifying the confidence around
the log mixing proportions. If a matrix, it must have dimensions \verb{N x (J + 1)}, with N being the number of observations and J being the number of
sources. The additional J + 1 column represents the unsampled source, making
the final row sums across the J + 1 columns total 1.
\item \code{m}: A numeric value specifying the sample size for each source. Used when
\code{sample_tracer} is \code{TRUE}.
\item \code{sample_tracer}: Logical. If \code{TRUE}, the model estimates uncertainty
around sampled source signatures.
\item \code{fix_unsampled}: Logical. If \code{TRUE}, the model fixes unsampled source signatures to the mean across all sampled sources.
\item \code{hierarchical}: Logical. If \code{TRUE}, the model includes a hierarchical grouping structure.
}
}
\examples{
library(mixmustr)
data(mixmustr_models)

# Example input data
synthetic_list <- list(
  df_stream_1 = data.frame(
    group = c("A", "A", "B", "B"),
    tracer1 = c(1.2, 1.3, 2.1, 2.2),
    tracer2 = c(3.4, 3.5, 4.6, 4.7)
  ),
  df_stream_2 = data.frame(
    group = c("A", "A", "B", "B"),
    source1 = c(0.6, 0.4, 0.7, 0.3),
    source2 = c(0.4, 0.6, 0.3, 0.7)
  )
)

mu_tab <- data.frame(
  source = c("source1", "source2"),
  tracer1 = c(1.25, 2.15),
  tracer2 = c(3.45, 4.65)
)

sig_tab <- data.frame(
  source = c("source1", "source2"),
  tracer1 = c(0.05, 0.1),
  tracer2 = c(0.15, 0.2)
)

# Example parameters
model_path <- tempfile(fileext = ".stan")
sigma_ln_rho <- 0.01
m <- 10
sample_tracer <- TRUE
fix_unsampled <- FALSE
hierarchical <- TRUE

# Call the function
input_list <- mixmustr_wrangle_input(
  data_streams_list = synthetic_list,
  mu_tab = mu_tab,
  sig_tab = sig_tab,
  model_path = model_path,
  sigma_ln_rho = sigma_ln_rho,
  m = m,
  sample_tracer = sample_tracer,
  fix_unsampled = fix_unsampled,
  hierarchical = hierarchical
)

}
\seealso{
\code{\link{run_mixmustr_models}}
}
