% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{reshape_ref_data}
\alias{reshape_ref_data}
\title{Calculates unsampled mixing proportions}
\usage{
reshape_ref_data(data_streams_list, target = "df_stream_2", order_ref)
}
\arguments{
\item{data_streams_list}{A list containing the input
data streams for the model. It should include two
data frames named \code{df_stream_1} and\code{df_stream_2}. See
details for the expected structure of these.}

\item{target}{The target element of \code{data_streams_list}, typically
\code{"df_stream_2"}.}

\item{order_ref}{A character vector specifying the correct order of the
source names.}
}
\value{
A matrix of N x (J + 1), with the + 1 representing the
mixing proportions from the unsampled source(s).
}
\description{
Calculates unsampled mixing proportions
}
\examples{
library(mixmustr)
reshape_ref_data(synthetic_df_convergent, target = "df_stream_2",
                 order_ref = tracer_parameters$mus$source)

}
