% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{run_mixmod}
\alias{run_mixmod}
\title{run_mixmod}
\usage{
run_mixmod(model_path, mod_name_suffix, ...)
}
\arguments{
\item{model_path}{A character string specifying the path to the Stan model
file.}

\item{mod_name_suffix}{A character string used to create a unique model name.}

\item{...}{Additional arguments passed to \code{\link[rstan]{stan}}.}
}
\description{
Wrapper to run mixture model in Stan.
}
\seealso{
\code{\link{run_mixmustr_models}}
}
