% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{run_mixmustr_models}
\alias{run_mixmustr_models}
\title{run_mixmustr_models}
\usage{
run_mixmustr_models(
  model_choices,
  data_streams_list,
  mu_tab,
  sig_tab = NULL,
  sigma_ln_rho,
  ...
)
}
\arguments{
\item{model_choices}{A data frame specifying the model configurations to
run. Each row should represent a model, with columns indicating the values
for \code{sample_tracer}, \code{fix_unsampled}, \code{hierarchical}, and \code{code_path}.}

\item{data_streams_list}{A list containing the input
data streams for the model. It should include two
data frames named \code{df_stream_1} and\code{df_stream_2}. See
details for the expected structure of these.}

\item{mu_tab}{A data frame containing the mean tracer
signatures for each source. The first column should be named \code{source}, and
the remaining columns should contain numeric tracer values.}

\item{sig_tab}{(Optional) A data frame containing the
standard deviations of tracer signatures for each source. The structure must
match \code{mu_tab}. Required if \code{sample_tracer} is \code{TRUE} or if \code{fix_unsampled}
is \code{FALSE} and \code{sample_tracer} is \code{FALSE}.}

\item{sigma_ln_rho}{A numeric value or matrix specifying the confidence
around the log mixing proportions. If a matrix, it must be of dimensions
N X J, with N being the number of observations and J being the number of
sources.}

\item{...}{Additional arguments passed to \code{\link[rstan]{stan}}, such as
\code{iter}, \code{chains}, or \code{control}.}
}
\value{
A list where each element corresponds to a model configuration,
containing the run time and model output.
}
\description{
Runs a set of Stan models.
}
\details{
The \code{run_mixmustr_models} function automates the process of running
multiple \code{mixmustr} models based on user-specified configurations. It builds
the Stan code for each model, prepares the input data, and runs the models
sequentially. The results are returned as a list, with each element
containing the run time and model output for a specific configuration. For
convenience, \code{mixmustr} provides the user with a built-in data frame
containing all of the potential allowed models \code{\link{mixmustr_models}}.
}
\examples{
\dontrun{
library(mixmustr)
mus <- tracer_parameters$mus
sigmas <- tracer_parameters$sigmas
model_fits <- run_mixmustr_models(
  mixmustr_models[1, ], synthetic_df_convergent, mus, sigmas,
  sigma_ln_rho = 0.1, iter = 1e4, warmup = 5e3, chains = 4, cores = 4
)
}

}
\seealso{
\code{\link{mixmustr_models}},
\code{\link{mixmustr_wrangle_input}}
}
