% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{run_mixmustr_models}
\alias{run_mixmustr_models}
\title{run_mixmustr_models}
\usage{
run_mixmustr_models(
  model_choices,
  data_streams_list,
  tracer_list,
  sigma_ln_rho,
  ...
)
}
\arguments{
\item{model_choices}{A data frame specifying the model configurations to
run. Each row should represent a model, with columns indicating the values
for \code{sample_tracer}, \code{fix_unsampled}, \code{hierarchical}, and \code{code_path}.}

\item{data_streams_list}{A list containing the input
data streams for the model. It should include two
data frames named \code{df_stream_1} and\code{df_stream_2}. See
details for the expected structure of these.}

\item{tracer_list}{A named list containing 1--3 data frames of same size,
one for the mean signatures, one for their standard deviations, and another
one for sample size. The second data frame is mandatory if \code{sample_tracer}
is \code{TRUE}, or if \code{fix_unsampled} is \code{FALSE} and \code{sample_tracer} is \code{FALSE}.
The third data frame is mandatory if \code{sample_tracer} is \code{TRUE}. See details
for exact structure of these data frames.}

\item{sigma_ln_rho}{A numeric value or matrix specifying the confidence
around the log mixing proportions. If a matrix, it must be of dimensions
N X J, with N being the number of observations and J being the number of
sources.}

\item{...}{Additional arguments passed to \code{\link[rstan]{stan}}, such as
\code{iter}, \code{chains}, or \code{control}.}
}
\value{
A list where each element corresponds to a model configuration,
containing the run time and model output.
}
\description{
Runs a set of Stan models.
}
\details{
The \code{run_mixmustr_models} function automates the process of running
multiple \code{mixmustr} models based on user-specified configurations. It builds
the Stan code for each model, prepares the input data, and runs the models
sequentially. The results are returned as a list, with each element
containing the run time and model output for a specific configuration. For
convenience, \code{mixmustr} provides the user with a built-in data frame
containing all of the potential allowed models \code{\link{mixmustr_models}}.
}
\examples{
\dontrun{
library(mixmustr)
# mixmustr_models[6, ] runs quickest
model_fits <- run_mixmustr_models(
  mixmustr_models[6, ], synthetic_df_convergent, tracer_parameters,
  sigma_ln_rho = 0.1, iter = 1e4, warmup = 5e3, chains = 4, cores = 4
)
}

}
\seealso{
\code{\link{mixmustr_models}},
\code{\link{mixmustr_wrangle_input}}
}
