% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_df_divergent.R
\docType{data}
\name{synthetic_df_divergent}
\alias{synthetic_df_divergent}
\title{Synthetic Divergent Dataset}
\format{
A list containing three data frames:
\describe{
\item{df_stream_1}{A data frame containing the simulated mixture data for the first stream, including tracer estimates calculated from the stable isotope and fatty acid data.}
\item{df_stream_2}{A data frame containing the simulated synthetic
proportions for the second stream, a column for each source.}
\item{stream_1_props}{A data frame containing the synthetic proportions for
the first stream, used as input for generating the mixture data. This is only
used for testing purposes}
}
}
\source{
The synthetic dataset was generated programmatically using the \code{MixMustR} package. The input data sources are:
\itemize{
\item Stable isotope data: See \code{bcs_si} documentation.
\item Fatty acid data: See \code{bcs_fa} documentation.
}
}
\usage{
synthetic_df_divergent
}
\description{
This dataset is a synthetic dataset generated to simulate mixture data for testing and validation purposes in the \code{MixMustR} package. It is anchored to empirical values of stable isotopes and fatty acids for a range of plant carbon sources in marine soils stable isotope data (\code{bcs_si}), fatty acid data (\code{bcs_fa}), and synthetic mixture proportions. \code{synthetic_df_divergent} exhibits great difference in the underlying mixing proportions between data streams 1 and 2.
}
\details{
The dataset was generated using the non-exported \code{make_mixture_data} function, which combines stable isotope data (\code{bcs_si}), fatty acid data (\code{bcs_fa}), and synthetic proportions (\code{stream_1_props} and \code{stream_2_props}) to produce simulated mixture data. The \code{truth_stream} parameter determines which stream's proportions are used as the "true" source contributions.
}
\examples{
library(MixMustR)
data(synthetic_df_divergent)
str(synthetic_df_divergent)
}
\seealso{
\code{\link{synthetic_df_convergent}}
}
\keyword{datasets}
